change 'type' prop to 'color'
+ A bit of housekeeping / cleanup (within `/lib/helpers`)

Using the `type` property on <Button/> conflicted w/ the HTML
attribute of the same name. It was decided that this custom prop
should then be renamed to `color`. As `color="link"` (formerly "text")
doesn't make sense, the `link` prop was added to Button.

So

    <Button type="link">Click here to return</Button>

now becomes

    <Button link>Click here to return</Button>

This allows further customization by allowing users to use the `color`
prop to style the color of link (though not recommended, as it'll generate
an extra css class for each variant).

Closes #1